progress 8

+++ app.js
@@ -14,7 +14,7 @@ env.config();
 
 const db = new pg.Client({
   user: process.env.PG_USER,
-  host: process.env.PG_host,
+  host: process.env.PG_HOST,
   database: process.env.PG_DATABASE,
   password: process.env.PG_PASSWORD,
   port: process.env.PG_PORT,
@@ -28,7 +28,7 @@ app.use(
     resave: false,
     saveUninitialized: true,
     cookie: {
-      maxAge: 1000 * 60 * 60,
+      maxAge: 1000 * 60 * 60 * 60,
     },
   })
 );
@@ -96,7 +96,7 @@ app.post("/signup", async (req, res) => {
 
 app.get("/home", isAuthenticated, async (req, res) => {
   let result = await db.query(
-    "SELECT users.first_name,users.last_name,posts.title,posts.detail,posts.created_at FROM users INNER JOIN posts ON users.id = posts.user_id ORDER BY created_at DESC"
+    "SELECT users.id, users.first_name,users.last_name,posts.title,posts.detail,posts.created_at FROM users INNER JOIN posts ON users.id = posts.user_id ORDER BY created_at DESC"
   );
   let posts = result.rows;
 
@@ -124,29 +124,63 @@ app.post("/compose", isAuthenticated, async (req, res) => {
 
 app.get("/profile", isAuthenticated, async (req, res) => {
   const user = req.user;
-  let result = await db.query(
+  const result = await db.query(
     "Select * FROM posts WHERE user_id=$1 ORDER BY created_at DESC",
     [user.id]
   );
-  let posts = result.rows;
+  const posts = result.rows;
+  const profileUser = req.user;
+
   res.render("profile", {
     posts: posts,
-    first: user.first_name,
-    last: user.last_name,
+    profileUser: profileUser,
+    currentUser: user,
   });
 });
 
-app.get("/delete/:id", isAuthenticated, async (req, res) => {
-  let id = req.params.id;
-  await db.query(`DELETE FROM posts WHERE id=$1`, [id]);
+app.get("/profile/:id", isAuthenticated, async (req, res) => {
+  const user = req.user;
+  const id = req.params.id;
+  try {
+    const userResult = await db.query("Select * FROM users WHERE id=$1 ", [id]);
+    const profileUser = userResult.rows[0];
+    if (!profileUser) {
+      return res.redirect("/home");
+    }
+
+    const postResult = await db.query(
+      "Select * FROM posts WHERE user_id=$1 ORDER BY created_at DESC",
+      [id]
+    );
+
+    const posts = postResult.rows;
+
+    res.render("profile", {
+      posts: posts,
+      profileUser: profileUser,
+      currentUser: user,
+    });
+  } catch (err) {
+    console.log(err);
+  }
+});
+
+app.post("/delete/:id", isAuthenticated, async (req, res) => {
+  const id = req.params.id;
+  const user = req.user;
+  await db.query(`DELETE FROM posts WHERE id=$1 And user_id=$2`, [id, user.id]);
   res.redirect("/profile");
 });
 
 app.post("/update/:id", isAuthenticated, async (req, res) => {
-  let id = req.params.id;
-
-  let detail = req.body.detail.trim();
-  await db.query(`UPDATE posts SET detail=$1 WHERE id=$2`, [detail, id]);
+  const id = req.params.id;
+  const user = req.user;
+  const detail = req.body.detail.trim();
+  await db.query(`UPDATE posts SET detail=$1 WHERE id=$2  And user_id=$3`, [
+    detail,
+    id,
+    user.id,
+  ]);
   res.redirect("/profile");
 });
